format text and dates

+++ app/src/main/java/com/example/krzdabrowski/myapplication/model/Event.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Event(
     @SerializedName("title") val name: String,
-    @SerializedName("event_date_unix") val date: Int,
+    @SerializedName("event_date_unix") val date: Long,
     @SerializedName("details") val info: String,
     @SerializedName("links") val urls: Map<String, String>
 )
